Some new conversions to double
git-svn-id: https://ferma.svn.sf.net/svnroot/ferma/branches/double-coords-migration@312 d51c77f7-620d-0410-a109-9a2f07eaa567

+++ FermaNext/src/TrussUnit/Truss.h
@@ -219,7 +219,7 @@ public:
         return findAdjoiningPivots( node, true );
     }
 
-    virtual N& createNode ( int x, int y )
+    virtual N& createNode ( double x, double y )
     {
         // Clean earlier desisted nodes/pivots
         suspendedClean();
@@ -241,7 +241,7 @@ public:
         // Just subsidiary catches
         QObject::connect( node, SIGNAL(onFixationChange ( Fixation )),
                                 SLOT(stateIsChanged()) );
-        QObject::connect( node, SIGNAL(onPositionChange ( int, int )),
+        QObject::connect( node, SIGNAL(onPositionChange(double, double)),
                                 SLOT(stateIsChanged()) );
 
         emit afterNodeCreation(*node);
